chore(wizard-refactor): extract small tools (show_options/show_upload/show_suggestions) + tool-exec types
Introduces wizard-tool-executor/_shared/types.ts (WizardToolResult +
WizardToolExec contract). Monolith re-exports WizardToolResult for
backwards-compat with chat/route.ts.

Each per-tool module exports a single `execute(input, userId, setupData)`
returning Omit<WizardToolResult, "tool_use_id"> — dispatcher injects
tool_use_id via spread. No behaviour change.

Dispatcher 11/11 + lint 17/17 + 33 sibling tests green (61 total).
