+++ apps/admin/lib/chat/wizard-tool-executor/tools/show_options.ts
@@ -1,0 +1,5 @@
+import type { WizardToolExec } from "../_shared/types";
+
+export async function execute(): Promise<WizardToolExec> {
+  return { content: `Panel displayed to user. Wait for their response.` };
+}
